Incomplete implementation, contains many errors.
Needs a step back on how preorder() works that does more to serve its
end-user than it currently does. Ideally, preorder is somehow consumed
by methods like reduce, traverse, map, etc as an option of how to
traverse the tree. Maybe it creates a new object from the tree that
implements an interface similar to (or an actual implementation of)
Iterable so that methods like reduce can just use the Iterable to get
the next node & the node's depth w/out knowing anything about how the
actual traversal works. This also would allow for other implementers of
the interface to be passed in that implement a different traversal
order, such as post-order or in-order.